add login with email and password and some modification to some components

+++ lib/components/AccountDetails/delete_account_button.dart
@@ -2,6 +2,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:modifai/components/ImageViewer/modifai_progress_indicator.dart';
+import 'package:modifai/components/confirmation_dialog.dart';
 
 import '../../screens/registration.dart';
 
@@ -17,17 +18,23 @@ class _DeleteAccountButtonState extends State<DeleteAccountButton> {
   Widget build(BuildContext context) {
     return TextButton(
       onPressed: () async {
-        if (FirebaseAuth.instance.currentUser != null) {
-          DialogUtils.modifAiProgressindicator();
-          await FirebaseAuth.instance.currentUser!.delete();
-          Get.offAll(() => const Registration());
-          Get.snackbar(
-            "Alert",
-            "Your account is deleted now, please register to access all the features",
-            colorText: Colors.white,
-            snackPosition: SnackPosition.BOTTOM,
-          );
-        }
+        Get.dialog(
+          ConfirmationDialog(
+            onConfirm: () async {
+              if (FirebaseAuth.instance.currentUser != null) {
+                DialogUtils.modifAiProgressindicator();
+                await FirebaseAuth.instance.currentUser!.delete();
+                Get.offAll(() => const Registration());
+                Get.snackbar(
+                  "Alert",
+                  "Your account is deleted now, please register to access all the features",
+                  colorText: Colors.white,
+                  snackPosition: SnackPosition.BOTTOM,
+                );
+              }
+            },
+          ),
+        );
       },
       child: const Text(
         'Delete Account',
